extract out separate component files
