+++ app/src/components/donor/home/FileUploadRow.tsx
@@ -1,0 +1,67 @@
+import { useId, useRef } from "react";
+import { Label } from "@/components/ui/label";
+
+export function FileUploadRow({
+  fileName,
+  onFile,
+  onClear,
+  showClear = false,
+}: {
+  fileName: string | null;
+  onFile: (name: string) => void;
+  onClear: () => void;
+  showClear?: boolean;
+}) {
+  const inputId = useId();
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  return (
+    <div className="flex flex-row items-center gap-4">
+      <Label
+        htmlFor={inputId}
+        className="flex shrink-0 cursor-pointer flex-col gap-0 text-base font-medium leading-tight text-gray-800"
+      >
+        <span>Attach</span>
+        <span>Receipt</span>
+      </Label>
+      <div className="flex min-w-0 flex-1 flex-col items-end gap-1">
+        <input
+          ref={inputRef}
+          id={inputId}
+          type="file"
+          accept="image/*,.pdf"
+          className="sr-only"
+          onChange={(e) => {
+            const f = e.target.files?.[0];
+            if (f) onFile(f.name);
+            e.target.value = "";
+          }}
+        />
+        <div className="flex items-center gap-2 w-full justify-end">
+          <button
+            type="button"
+            className="inline-flex min-h-10 items-center justify-center rounded-xl bg-kfk-blue px-8 py-2 font-gaegu text-[18px] font-bold text-white transition-colors hover:bg-kfk-blue/80"
+            onClick={() => inputRef.current?.click()}
+          >
+            Upload
+          </button>
+          {showClear && fileName && (
+            <button
+              type="button"
+              onClick={onClear}
+              className="text-gray-500 hover:text-gray-700 text-xl font-medium leading-none"
+              aria-label="Remove file"
+            >
+              ×
+            </button>
+          )}
+        </div>
+        {fileName && (
+          <p className="max-w-[220px] truncate text-right text-xs text-gray-500">
+            {fileName}
+          </p>
+        )}
+      </div>
+    </div>
+  );
+}
